creates database on server start if it doesn't already exist
