--- db_setup.js
@@ -1,25 +1,0 @@
-var sqlite3 = require('sqlite3').verbose();
-var db = new sqlite3.Database('mydb.db');
-var check;
-
-exports.sayhi = sayhi;
-
-function sayhi() {
-  console.log("Hello from the other side");
-}
-
-db.serialize(function() {
-
-  db.run("CREATE TABLE if not exists user_info (info TEXT)");
-  var stmt = db.prepare("INSERT INTO user_info VALUES (?)");
-  for (var i = 0; i < 10; i++) {
-      stmt.run("Ipsum " + i);
-  }
-  stmt.finalize();
-
-  db.each("SELECT rowid AS id, info FROM user_info", function(err, row) {
-      console.log(row.id + ": " + row.info);
-  });
-});
-
-db.close();
